move server code to an importable package
Commit https://github.com/cloudflare/redoctober/commit/6f8424ad38982c92d64b9c70fb1e2e823aeb9f18
added an public function so we can import redoctober's NewServer
function in external test packages to create an RO server without having
to actually install the binary in test environments.

This used to work until https://github.com/golang/go/commit/0f06d0a051714d14b923b0a9164ab1b3f463aa74,
which makes it impossible to import main package in external packages.

This change moves `NewServer` and its related code to a non-main
package so other packages can still import it in tests or any other
places.
